added all relationships

+++ app/src/main/java/ch/hearc/ezworkout/database/LogBookPage.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "LogBookPage")
 data class LogBookPage(
     @PrimaryKey(autoGenerate = true)
-    val id: Int,
-    val trainingPlanId: Int
+    val id: Long,
+    val trainingPlanId: Long
 )
